refactoring UserInfo component
